Split ChatAgent into MainAgent and SubAgent
After removing Administrator/RouterAgent/handoff, the one remaining
ChatAgent class conflated two behaviourally different things: the
user-facing agent that streams responses with identity, and the
task-only worker that a main agent delegates to. Split them so the
two roles are explicit:

- BaseAgent (abstract): shared tool-approval + MCP fallback plumbing.
- MainAgent: owns identity (voice, avatar, character, preferred_name),
  streams StreamChunkEvents to the user, emits ContextBreakdownEvent
  per turn, accepts handler-injected SubAgentTool adapters.
- SubAgent: no identity, no streaming to the frontend, runs its own
  LLM + tool-loop internally, returns a single string to its caller
  via SubAgent.execute(task, context).
- SubAgentTool: thin BaseTool adapter that wraps a SubAgent so a
  MainAgent's LLM can call it.

tools/subagent.py moves into agents/sub.py since the adapter is agent
infrastructure, not a general tool. Handlers instantiate MainAgent
directly and inject SubAgentTool-wrapped SubAgents as extra tools.
74 unit tests pass.
