+++ kurisuassistant/agents/base.py
@@ -1,16 +1,20 @@
-"""Base agent class."""
+"""Base agent plumbing: shared dataclasses, tool execution, MCP fallback.
+
+Concrete agents live in ``main.py`` (MainAgent) and ``sub.py`` (SubAgent).
+BaseAgent itself has no ``process()`` / ``execute()`` — the two concrete
+agents have different calling conventions (streaming to the user vs.
+returning a single string), so the interface diverges.
+"""
 
 import asyncio
 import json
 import logging
+from abc import ABC
 from dataclasses import dataclass, field
-from typing import AsyncGenerator, Callable, Coroutine, Dict, List, Any, Optional, TYPE_CHECKING
+from typing import Callable, Coroutine, Dict, List, Any, Optional, TYPE_CHECKING
 
 from kurisuassistant.websocket.events import (
-    StreamChunkEvent,
     ToolApprovalRequestEvent,
-    ToolApprovalResponseEvent,
-    ContextBreakdownEvent,
 )
 from kurisuassistant.tools import ToolRegistry
 
@@ -46,31 +50,31 @@ async def async_iterate(sync_iterator):
 
 @dataclass
 class AgentConfig:
-    """Configuration for an agent with identity and capabilities.
+    """Configuration shared by MainAgent and SubAgent.
 
-    Main agents (agent_type='main') have personality: voice, avatar, character config.
-    Sub-agents (agent_type='sub') are tools with no personality.
+    MainAgent requires identity fields (voice_reference, avatar_uuid,
+    character_config, preferred_name, trigger_word). SubAgent ignores them.
     """
     id: Optional[int] = None
-    name: str = ""  # Agent/character name
-    description: str = ""  # Short description for routing LLM
-    system_prompt: str = ""  # Role + personality instructions
+    name: str = ""
+    description: str = ""
+    system_prompt: str = ""
 
-    # Identity (merged from former Persona model)
+    # Identity (MainAgent only)
     voice_reference: Optional[str] = None
     avatar_uuid: Optional[str] = None
-    character_config: Optional[Dict] = None  # Animation tree config
-    preferred_name: Optional[str] = None  # How user wants to be called
+    character_config: Optional[Dict] = None
+    preferred_name: Optional[str] = None
 
     # Inference config
     model_name: Optional[str] = None
     provider_type: str = "ollama"
-    available_tools: Optional[List[str]] = None  # None = all tools, list = only these
+    available_tools: Optional[List[str]] = None  # None = all tools
     think: bool = False
     use_deferred_tools: bool = False
 
-    # Agent type and state
-    agent_type: str = "main"  # 'main' (personality) or 'sub' (tool)
+    # State
+    agent_type: str = "main"  # 'main' or 'sub'
     memory: Optional[str] = None
     memory_enabled: bool = True
     enabled: bool = True
@@ -88,14 +92,14 @@ class AgentContext:
     available_agents: List[AgentConfig] = field(default_factory=list)
     user_system_prompt: str = ""
     preferred_name: str = ""
-    api_url: Optional[str] = None  # User-specific Ollama endpoint
-    gemini_api_key: Optional[str] = None  # User-specific Gemini API key
-    nvidia_api_key: Optional[str] = None  # User-specific NVIDIA NIM API key
+    api_url: Optional[str] = None
+    gemini_api_key: Optional[str] = None
+    nvidia_api_key: Optional[str] = None
     client_tools: List[Dict] = field(default_factory=list)
     client_tool_callback: Optional[Callable[[str, Dict], Coroutine[Any, Any, str]]] = None
-    images: Optional[List[str]] = None  # base64 images for current user message
-    context_size: Optional[int] = None  # Ollama num_ctx override
-    compacted_context: str = ""  # Rolling conversation summary (short-term memory)
+    images: Optional[List[str]] = None
+    context_size: Optional[int] = None
+    compacted_context: str = ""
 
 
 @dataclass
@@ -107,10 +111,8 @@ class ToolResult:
 
     @staticmethod
     def _detect_error(content: str) -> bool:
-        """Check if tool content indicates an error."""
         import json as _json
         stripped = content.strip()
-        # Detect JSON error responses: {"error": "..."}
         if stripped.startswith("{"):
             try:
                 parsed = _json.loads(stripped)
@@ -118,20 +120,17 @@ class ToolResult:
                     return True
             except (_json.JSONDecodeError, ValueError):
                 pass
-        # Detect wrapped client/MCP tool errors
         if stripped.startswith("Client tool error:") or stripped.startswith("MCP client not available"):
             return True
         return False
 
     @staticmethod
     def _detect_denied(content: str) -> bool:
-        """Check if tool content indicates user denial."""
         lc = content.lower()
         return "denied by user" in lc or "denied by the user" in lc
 
     @staticmethod
     def from_content(content: str, **kwargs) -> "ToolResult":
-        """Create a ToolResult, auto-detecting error/denied status from content."""
         if ToolResult._detect_denied(content):
             status = "denied"
         elif ToolResult._detect_error(content):
@@ -141,21 +140,18 @@ class ToolResult:
         return ToolResult(content=content, status=status, **kwargs)
 
 
-class ChatAgent:
-    """Conversational agent — streams an LLM, loops on tool calls, yields context breakdowns.
+class BaseAgent(ABC):
+    """Abstract base: shared tool-approval plumbing for Main and Sub agents.
 
-    Exposes tool approval through the handler (frontend decides via tool_policies),
-    injects agent identity / memory / skills / compacted context into the system
-    prompt, and emits ``ContextBreakdownEvent`` at the start of each LLM round.
+    Subclasses implement their own top-level entry point (``process`` for
+    streaming main agents, ``execute`` for one-shot sub agents). The
+    tool-approval flow and MCP fallback live here because they're identical
+    for both.
     """
 
     def __init__(self, config: AgentConfig, tool_registry: ToolRegistry):
         self.config = config
         self.tool_registry = tool_registry
-        # Per-turn tracking for accurate raw_input/raw_output
-        self.turn_data: List[Dict[str, Any]] = []
-        self.context_breakdown: Dict[str, Any] = {}
-        self.loaded_skills: List[str] = []
 
     async def execute_tool(
         self,
@@ -166,16 +162,8 @@ class ChatAgent:
         """Execute a tool after requesting approval from frontend.
 
         All tool calls go through frontend for permission check. Frontend
-        decides whether to auto-approve based on user's policies or show
-        approval dialog.
-
-        Args:
-            tool_name: Name of the tool to execute
-            tool_args: Arguments for the tool
-            context: Agent context
-
-        Returns:
-            ToolResult with content and optional images
+        decides whether to auto-approve based on the user's policies or
+        show an approval dialog.
         """
         import json as _json
 
@@ -188,7 +176,6 @@ class ChatAgent:
             return await self.execute_tool(inner_name, inner_args, context)
 
         # Deferred tools: handle list_tools / search_tools / get_tool_schema via proxy
-        # These are meta-tools that don't require approval
         proxy = getattr(self, "_deferred_proxy", None)
         if proxy and tool_name == "list_tools":
             page = tool_args.get("page", 1)
@@ -203,39 +190,33 @@ class ChatAgent:
             content = await proxy.get_tool_schema(name)
             return ToolResult(content=content)
 
-        # Determine tool location and get tool object if available
         tool = self.tool_registry.get(tool_name)
         execution_location = "backend"
 
-        # Check extra_tools if not found in registry (sub-agent tools, handoff tool)
+        # Check extra_tools if not found in registry (e.g. SubAgentTool adapters)
         if tool is None and hasattr(self, 'extra_tools') and self.extra_tools:
             for extra_tool in self.extra_tools:
                 if extra_tool.name == tool_name:
                     tool = extra_tool
                     break
 
-        # Check if this is a client-side tool
         client_tool_names = {
             t.get("function", {}).get("name", "") for t in (context.client_tools or [])
         }
         if tool is None and tool_name in client_tool_names:
             execution_location = "frontend"
 
-        # Enforce tool access: built-in tools always available,
-        # available_tools allowlist restricts non-built-in tools
         if self.config.available_tools is not None and tool_name not in self.config.available_tools:
             if not (tool and tool.built_in):
                 logger.warning(f"Agent '{self.config.name}' tried to use unavailable tool: {tool_name}")
                 return ToolResult(content=f"Tool not available: {tool_name}", status="error")
 
-        # Get human-readable description
         if tool:
             description = tool.describe_call(tool_args)
         else:
             description = f"Execute {tool_name} with args: {tool_args}"
 
         # ALWAYS request approval from frontend
-        # Frontend decides whether to auto-approve or show dialog based on user's policies
         if context.handler:
             approval_request = ToolApprovalRequestEvent(
                 tool_name=tool_name,
@@ -251,36 +232,21 @@ class ChatAgent:
             if not response.approved:
                 return ToolResult(content=f"Tool execution denied by user: {tool_name}", status="denied")
 
-            # Use modified args if provided
             if response.modified_args:
                 tool_args = response.modified_args
 
-        # Copy args before injecting internal keys to avoid mutating the
-        # caller's dict (which may be referenced by tool_calls in messages
-        # sent back to the LLM).
         exec_args = dict(tool_args)
 
-        # Inject conversation_id for context-aware tools
         if context.conversation_id:
             exec_args["conversation_id"] = context.conversation_id
-
-        # Inject user_id for user-scoped tools (e.g. skill lookup)
         if context.user_id:
             exec_args["user_id"] = context.user_id
-
-        # Inject agent_id for agent-scoped tools (e.g. knowledge graph)
         exec_args["agent_id"] = self.config.id
-
-        # Inject handler for tools that need WebSocket access (e.g. media player)
         if context.handler:
             exec_args["_handler"] = context.handler
-
-        # Inject full context for sub-agent tools
         exec_args["_context"] = context
 
-        # Execute based on location
         if execution_location == "frontend":
-            # Client-side tool - execute via callback
             if context.client_tool_callback:
                 try:
                     result = await context.client_tool_callback(tool_name, tool_args)
@@ -291,7 +257,6 @@ class ChatAgent:
             else:
                 return ToolResult(content=f"No client tool callback for: {tool_name}", status="error")
 
-        # Backend execution
         if tool:
             try:
                 result = await tool.execute(exec_args)
@@ -300,7 +265,6 @@ class ChatAgent:
                 logger.error(f"Tool execution failed: {e}", exc_info=True)
                 return ToolResult(content=f"Tool execution failed: {e}", status="error")
         elif context.user_id:
-            # Try server-side MCP tool
             return await self._execute_mcp_tool(tool_name, tool_args, context)
         else:
             return ToolResult(content=f"Unknown tool: {tool_name}", status="error")
@@ -311,28 +275,14 @@ class ChatAgent:
         tool_args: Dict[str, Any],
         context: AgentContext,
     ) -> ToolResult:
-        """Execute a server-side MCP tool.
-
-        Args:
-            tool_name: MCP tool name
-            tool_args: Tool arguments
-            context: Agent context
-
-        Returns:
-            ToolResult with content and optional images
-        """
+        """Execute a server-side MCP tool."""
         try:
             from kurisuassistant.mcp_tools.orchestrator import get_user_orchestrator
             orchestrator = get_user_orchestrator(context.user_id)
 
-            # Ensure tool-to-client mapping is loaded (uses cache if fresh)
             await orchestrator.get_tools()
-
-            # Don't inject internal context (user_id, agent_id, etc.) into
-            # external MCP tool args — external servers don't expect them.
             mcp_args = dict(tool_args)
 
-            # Create a mock tool call object
             class MockToolCall:
                 class Function:
                     def __init__(self, name, args):
@@ -345,7 +295,7 @@ class ChatAgent:
             mock_call = MockToolCall(tool_name, mcp_args)
             results = await orchestrator.execute_tool_calls(
                 [mock_call],
-                conversation_id=context.conversation_id
+                conversation_id=context.conversation_id,
             )
 
             if results:
@@ -358,435 +308,3 @@ class ChatAgent:
             logger.warning(f"Server MCP tool execution failed for '{tool_name}': {e}")
 
         return ToolResult(content=f"Unknown tool: {tool_name}", status="error")
-
-    def _prepare_messages(
-        self,
-        messages: List[Dict],
-        context: AgentContext,
-    ) -> List[Dict]:
-        """Prepare messages for LLM by filtering and enriching.
-
-        - Builds a unified system prompt with agent persona + user preferences
-        - Filters out system messages from conversation history (already incorporated)
-        - Tracks token counts for context breakdown
-        """
-        import datetime
-
-        # Build agent descriptions for context
-        agent_descriptions = []
-        for agent in context.available_agents:
-            if agent.name == self.config.name:
-                continue  # Skip self
-            desc = agent.system_prompt[:150] if agent.system_prompt else "General assistant"
-            agent_descriptions.append(f"- {agent.name}: {desc}")
-
-        # Track individual component tokens for breakdown
-        breakdown = {
-            "system_prompt_tokens": 0,
-            "memory_tokens": 0,
-            "compacted_context_tokens": 0,
-            "skills_tokens": 0,
-            "tools_guidance_tokens": 0,
-            "other_agents_tokens": 0,
-            "message_history_tokens": 0,
-            "message_count": 0,
-        }
-
-        # Build unified system prompt: agent identity + user preferences
-        system_parts = []
-
-        # Base identity + system prompt
-        base_prompt = f"You are {self.config.name}."
-        if self.config.system_prompt:
-            base_prompt += "\n\n" + self.config.system_prompt
-        if context.user_system_prompt:
-            base_prompt += "\n\n" + context.user_system_prompt
-        preferred_name = self.config.preferred_name or context.preferred_name
-        if preferred_name:
-            base_prompt += f"\n\nThe user prefers to be called: {preferred_name}"
-        base_prompt += f"\n\nCurrent time: {datetime.datetime.utcnow().isoformat()}"
-        system_parts.append(base_prompt)
-        breakdown["system_prompt_tokens"] = estimate_tokens(base_prompt)
-
-        # Skills section
-        skill_names = []
-        if context.user_id:
-            from kurisuassistant.tools.skills import get_skill_names_for_user
-            skill_names = get_skill_names_for_user(context.user_id)
-            if skill_names:
-                skills_text = (
-                    "## Skills\n"
-                    "You have the following skills: " + ", ".join(skill_names) + ".\n"
-                    "Skills contain detailed instructions on HOW to perform specific tasks. "
-                    "You MUST call get_skill_instructions to load the relevant skill's instructions BEFORE "
-                    "attempting any task that matches a skill name. Do NOT guess or improvise — "
-                    "always read the skill first and follow its instructions exactly."
-                )
-                system_parts.append(skills_text)
-                breakdown["skills_tokens"] = estimate_tokens(skills_text)
-        self.loaded_skills = skill_names
-
-        # Deferred tools guidance
-        if self.config.use_deferred_tools:
-            tools_guidance = (
-                "## Tool Usage\n"
-                "You have access to tools through a discovery system. Use these functions:\n"
-                "1. list_tools(page?) — Browse available tools (name + description, paginated)\n"
-                "2. search_tools(query) — Search tools by keyword in name or description\n"
-                "3. get_tool_schema(name) — Get a tool's full parameter schema before calling it\n"
-                "4. call_tool(name, arguments) — Execute a tool\n\n"
-                "Workflow: list_tools or search_tools → get_tool_schema → call_tool.\n"
-                "You may skip discovery if you already know the tool name from context or a previous turn."
-            )
-            system_parts.append(tools_guidance)
-            breakdown["tools_guidance_tokens"] = estimate_tokens(tools_guidance)
-
-        # Agent memory
-        if self.config.memory_enabled and self.config.memory:
-            memory_text = "Your memory:\n" + self.config.memory
-            system_parts.append(memory_text)
-            breakdown["memory_tokens"] = estimate_tokens(memory_text)
-
-        # Compacted conversation context
-        if context.compacted_context:
-            compacted_text = "Conversation context:\n" + context.compacted_context
-            system_parts.append(compacted_text)
-            breakdown["compacted_context_tokens"] = estimate_tokens(compacted_text)
-
-        # Other agents
-        if agent_descriptions:
-            other_agents_text = (
-                "Other agents in this conversation:\n"
-                + "\n".join(agent_descriptions)
-                + "\n\nYou may see messages from these agents. "
-                "Just focus on your own response — "
-                "do not direct others to speak, ask them to chime in, "
-                "or manage the conversation flow. A separate system handles turn-taking."
-            )
-            system_parts.append(other_agents_text)
-            breakdown["other_agents_tokens"] = estimate_tokens(other_agents_text)
-
-        prepared = []
-        prepared.append({"role": "system", "content": "\n\n".join(system_parts)})
-
-        # Process message history
-        msg_count = 0
-        msg_tokens = 0
-        for msg in messages:
-            role = msg.get("role", "user")
-
-            # Skip system messages — already incorporated into agent's system prompt
-            if role == "system":
-                continue
-
-            content = msg.get("content", "")
-
-            # Preserve natural turn structure: user/assistant/tool keep their roles
-            if role == "assistant":
-                chat_role = "assistant"
-            elif role == "tool":
-                chat_role = "tool"
-            else:
-                chat_role = "user"
-
-            entry = {"role": chat_role, "content": content}
-            if chat_role == "assistant" and "thinking" in msg:
-                entry["thinking"] = msg["thinking"]
-            prepared.append(entry)
-
-            msg_count += 1
-            msg_tokens += estimate_tokens(content)
-            if "thinking" in msg:
-                msg_tokens += estimate_tokens(msg["thinking"])
-
-        breakdown["message_history_tokens"] = msg_tokens
-        breakdown["message_count"] = msg_count
-
-        # Store breakdown for later access
-        self.context_breakdown = breakdown
-
-        return prepared
-
-    async def process(
-        self,
-        messages: List[Dict],
-        context: AgentContext,
-    ) -> AsyncGenerator[StreamChunkEvent | ContextBreakdownEvent, None]:
-        """Process messages with LLM, looping on tool calls.
-
-        Flow: LLM call → stream response → if tool_calls, execute tools,
-        append results to messages, and call LLM again (up to 10 rounds).
-
-        Yields ContextBreakdownEvent at the start of each LLM turn with
-        detailed token counts and loaded tools/skills.
-        """
-        import json
-        from kurisuassistant.models.llm import create_llm_provider
-
-        # Determine model and provider
-        model = self.config.model_name or context.model_name
-        provider_type = self.config.provider_type or "ollama"
-
-        # Select API key based on provider
-        api_key = None
-        if provider_type == "gemini":
-            api_key = context.gemini_api_key
-        elif provider_type == "nvidia":
-            api_key = context.nvidia_api_key
-
-        llm = create_llm_provider(
-            provider_type,
-            api_url=context.api_url,
-            api_key=api_key,
-        )
-
-        messages = self._prepare_messages(messages, context)
-
-        # Expose prepared messages for raw_input logging
-        self.last_prepared_messages = messages
-
-        # Build tool schemas — deferred (meta-tools) or flat (all schemas)
-        allowed = set(self.config.available_tools) if self.config.available_tools is not None else None
-
-        if self.config.use_deferred_tools:
-            from kurisuassistant.tools.deferred import create_deferred_tools, META_TOOL_NAMES
-            proxy, meta_tools = create_deferred_tools(
-                tool_registry=self.tool_registry,
-                available_tools=allowed,
-                user_id=context.user_id,
-                client_tools=context.client_tools or [],
-            )
-            self._deferred_proxy = proxy
-            tool_schemas = [mt.get_schema() for mt in meta_tools]
-            # Include built-in tools directly in the tools param (not behind deferred discovery)
-            for tool in self.tool_registry._tools.values():
-                if tool.built_in and tool.name not in META_TOOL_NAMES:
-                    tool_schemas.append(tool.get_schema())
-        else:
-            self._deferred_proxy = None
-            tool_schemas = self.tool_registry.get_schemas(allowed)
-
-            # Add user's server-side MCP tools (filtered by allowlist)
-            if context.user_id:
-                try:
-                    from kurisuassistant.mcp_tools.orchestrator import get_user_orchestrator
-                    mcp_tools = await get_user_orchestrator(context.user_id).get_tools()
-                    if allowed is not None:
-                        mcp_tools = [t for t in mcp_tools if t.get("function", {}).get("name") in allowed]
-                    tool_schemas.extend(mcp_tools)
-                except Exception as e:
-                    logger.warning(f"Failed to load MCP tools for user {context.user_id}: {e}")
-
-            # Add client-side tools (filtered by allowlist)
-            if context.client_tools:
-                client_tools = context.client_tools
-                if allowed is not None:
-                    client_tools = [t for t in client_tools if t.get("function", {}).get("name") in allowed]
-                tool_schemas.extend(client_tools)
-
-        # Add extra tools (sub-agent tools, handoff tool) - injected by handler
-        if hasattr(self, 'extra_tools') and self.extra_tools:
-            for extra_tool in self.extra_tools:
-                tool_schemas.append(extra_tool.get_schema())
-
-        # Compute tool schema tokens
-        tool_schemas_json = json.dumps(tool_schemas, ensure_ascii=False)
-        tool_schemas_tokens = estimate_tokens(tool_schemas_json)
-        loaded_tool_names = [
-            t.get("function", {}).get("name", "unknown")
-            for t in tool_schemas
-        ]
-
-        # Attach base64 images to the last user message for vision models
-        if context.images:
-            for msg in reversed(messages):
-                if msg.get("role") == "user":
-                    msg["images"] = context.images
-                    break
-
-        # Deferred tools need extra rounds for discovery (list → schema → call)
-        MAX_TOOL_ROUNDS = 25 if self.config.use_deferred_tools else 10
-
-        # Context limit from settings or default
-        context_limit = context.context_size or 8192
-
-        # Reset turn data for this process() call
-        self.turn_data = []
-
-        try:
-            for turn in range(MAX_TOOL_ROUNDS):
-                # Snapshot raw input BEFORE LLM call
-                raw_input_messages = [dict(m) for m in messages]
-                self.last_prepared_messages = raw_input_messages
-
-                # Update message history tokens in breakdown (may have grown from tool results)
-                msg_tokens = 0
-                msg_count = 0
-                for msg in messages[1:]:  # Skip system message (already counted)
-                    msg_tokens += estimate_tokens(msg.get("content", ""))
-                    if "thinking" in msg:
-                        msg_tokens += estimate_tokens(msg["thinking"])
-                    msg_count += 1
-                self.context_breakdown["message_history_tokens"] = msg_tokens
-                self.context_breakdown["message_count"] = msg_count
-
-                # Calculate total tokens
-                total_tokens = (
-                    self.context_breakdown.get("system_prompt_tokens", 0) +
-                    self.context_breakdown.get("memory_tokens", 0) +
-                    self.context_breakdown.get("compacted_context_tokens", 0) +
-                    self.context_breakdown.get("skills_tokens", 0) +
-                    self.context_breakdown.get("tools_guidance_tokens", 0) +
-                    self.context_breakdown.get("other_agents_tokens", 0) +
-                    msg_tokens +
-                    tool_schemas_tokens
-                )
-
-                # Yield context breakdown at start of each turn
-                yield ContextBreakdownEvent(
-                    conversation_id=context.conversation_id,
-                    frame_id=context.frame_id,
-                    turn=turn,
-                    system_prompt_tokens=self.context_breakdown.get("system_prompt_tokens", 0),
-                    memory_tokens=self.context_breakdown.get("memory_tokens", 0),
-                    compacted_context_tokens=self.context_breakdown.get("compacted_context_tokens", 0),
-                    skills_tokens=self.context_breakdown.get("skills_tokens", 0),
-                    tools_guidance_tokens=self.context_breakdown.get("tools_guidance_tokens", 0),
-                    other_agents_tokens=self.context_breakdown.get("other_agents_tokens", 0),
-                    message_history_tokens=msg_tokens,
-                    message_count=msg_count,
-                    tool_schemas_tokens=tool_schemas_tokens,
-                    tool_count=len(tool_schemas),
-                    total_tokens=total_tokens,
-                    context_limit=context_limit,
-                    loaded_tools=loaded_tool_names,
-                    loaded_skills=self.loaded_skills,
-                )
-
-                stream = llm.chat(
-                    model=model,
-                    messages=messages,
-                    tools=tool_schemas if tool_schemas else [],
-                    stream=True,
-                    think=self.config.think,
-                    options={"num_ctx": context.context_size or 8192},
-                )
-
-                # Accumulate full response to detect tool calls and build follow-up messages
-                full_content = ""
-                full_thinking = ""
-                all_tool_calls = []
-
-                async for chunk in async_iterate(stream):
-                    msg = chunk.message
-
-                    # Handle thinking
-                    thinking = getattr(msg, 'thinking', None)
-                    if thinking:
-                        full_thinking += thinking
-                        yield StreamChunkEvent(
-                            content="",
-                            thinking=thinking,
-                            role="assistant",
-                            agent_id=self.config.id,
-                            name=self.config.name,
-                            conversation_id=context.conversation_id,
-                            frame_id=context.frame_id,
-                            model_name=model,
-                            provider_type=self.config.provider_type,
-                        )
-
-                    # Handle content
-                    if msg.content:
-                        full_content += msg.content
-                        yield StreamChunkEvent(
-                            content=msg.content,
-                            role="assistant",
-                            agent_id=self.config.id,
-                            name=self.config.name,
-                            conversation_id=context.conversation_id,
-                            frame_id=context.frame_id,
-                            model_name=model,
-                            provider_type=self.config.provider_type,
-                        )
-
-                    # Collect tool calls (typically arrive in final chunk(s))
-                    if hasattr(msg, 'tool_calls') and msg.tool_calls:
-                        all_tool_calls.extend(msg.tool_calls)
-
-                # No tool calls — LLM gave a final answer, done
-                if not all_tool_calls:
-                    break
-
-                # Append assistant message (with tool_calls) so Ollama sees the call context
-                assistant_msg = {"role": "assistant", "content": full_content}
-                if full_thinking:
-                    assistant_msg["thinking"] = full_thinking
-                tc_list = []
-                for tc in all_tool_calls:
-                    args = tc.function.arguments
-                    if isinstance(args, str):
-                        args = json.loads(args)
-                    tc_list.append({"function": {"name": tc.function.name, "arguments": args}})
-                assistant_msg["tool_calls"] = tc_list
-                messages.append(assistant_msg)
-
-                # Execute each tool and feed results back
-                tool_denied = False
-                for tc in all_tool_calls:
-                    tool_name = tc.function.name
-                    tool_args = tc.function.arguments
-                    if isinstance(tool_args, str):
-                        tool_args = json.loads(tool_args)
-
-                    result = await self.execute_tool(tool_name, tool_args, context)
-
-                    # For deferred call_tool, show the inner tool name to the client
-                    display_name = tool_name
-                    display_args = tool_args
-                    if tool_name == "call_tool":
-                        display_name = tool_args.get("name", "call_tool")
-                        display_args = tool_args.get("arguments", {})
-
-                    # Yield tool result (name=tool, no agent_id — tools aren't agents)
-                    yield StreamChunkEvent(
-                        content=result.content,
-                        role="tool",
-                        agent_id=None,
-                        name=display_name,
-                        conversation_id=context.conversation_id,
-                        frame_id=context.frame_id,
-                        tool_args=display_args,
-                        tool_status=result.status,
-                        images=result.images or None,
-                    )
-
-                    # Append tool result for next LLM round
-                    messages.append({"role": "tool", "content": result.content})
-
-                    # If tool was denied, stop executing remaining tools
-                    if result.status == "denied":
-                        tool_denied = True
-                        break
-
-                    # If handoff_to was called, stop immediately — don't loop back to LLM
-                    if display_name == "handoff_to":
-                        tool_denied = True  # Reuse flag to break outer loop
-                        break
-
-                # If user denied a tool or handoff_to was called, stop the agent's tool loop
-                if tool_denied:
-                    break
-
-                # Loop continues — LLM will see tool results and can call more tools or answer
-
-        except Exception as e:
-            logger.error(f"Agent processing failed: {e}", exc_info=True)
-            yield StreamChunkEvent(
-                content=f"Error: {e}",
-                role="assistant",
-                agent_id=self.config.id,
-                name=self.config.name,
-                conversation_id=context.conversation_id,
-                frame_id=context.frame_id,
-            )
